+++ Code/app/src/main/java/com/smarttrainer/smarttrainer/models/MotionJudge.java
@@ -6,7 +6,7 @@ import java.util.List;
  * Created by fan on 3/12/16.
  */
 public interface MotionJudge {
-    String judgeMotion(List<double[]> sensorRawData, int formID);
+    String judgeMotion(List<float[]> sensorRawData, int formID);
     void reset();
     int getCount();
 }
